requirements.txt
asgiref==3.7.2
Django==5.0.1
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==10.2.0
sqlparse==0.4.4
tzdata==2023.4
